student page done with some enrollements bugs

+++ Java/assigment2/src/main/java/com/assigment2/model/repositories/StudentRepository.java
@@ -31,7 +31,7 @@ public class StudentRepository {
 		return students;
 	}
 
-	public Student getById(Integer id) throws DatabaseAccesException {
+	public Student getById(Long id) throws DatabaseAccesException {
 		Session session = sessionFactory.openSession();
 		Student element = null;
 		try {
